requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib>=1.2.0
pyyaml>=6.0
streamlit>=1.25.0
plotly>=5.14.0
pydantic>=2.0.0
python-dotenv>=1.0.0